# swagger.yaml
{"openapi":"3.1.0","info":{"title":"Impacto Manager","description":"Sistema de gerenciamento de escola de esportes","license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Professor","description":"Endpoint para gerenciamento de professores"},{"name":"Aluno","description":"Endpoint para gerenciamento de alunos"},{"name":"Pagamento","description":"Gerenciamento do valor da mensalidade a ser paga pelos alunos"},{"name":"Turma","description":"Endpoint para gerenciamento de turmas"}],"paths":{"/api/turma":{"put":{"tags":["Turma"],"summary":"Update","description":"Updates an existing item","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Turma"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Turma"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"post":{"tags":["Turma"],"summary":"Create new","description":"Creates a new item","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Turma"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Turma"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/professor":{"put":{"tags":["Professor"],"summary":"Update","description":"Updates an existing item","operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Professor"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Professor"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"post":{"tags":["Professor"],"summary":"Create new","description":"Creates a new item","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Professor"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Professor"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/pagamento":{"get":{"tags":["Pagamento"],"summary":"List all","description":"Returns a list of all items","operationId":"findAll","responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Pagamento"}}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"put":{"tags":["Pagamento"],"summary":"Update","description":"Updates an existing item","operationId":"update_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pagamento"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pagamento"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"post":{"tags":["Pagamento"],"summary":"Create new","description":"Creates a new item","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pagamento"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pagamento"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"delete":{"tags":["Pagamento"],"summary":"Delete","description":"Deletes an item","operationId":"deleteById","parameters":[],"responses":{"200":{"description":"Operação realizada com sucesso"},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/aluno":{"get":{"tags":["Aluno"],"operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlunoDto"}}}}}}},"put":{"tags":["Aluno"],"summary":"Update","description":"Updates an existing item","operationId":"update_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Aluno"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Aluno"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"post":{"tags":["Aluno"],"summary":"Create new","description":"Creates a new item","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Aluno"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Aluno"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/user/register":{"post":{"tags":["user-controller"],"summary":"Create new","description":"Creates a new item","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/login":{"post":{"tags":["authentication-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/private":{"get":{"tags":["private-controller"],"operationId":"GetMessage","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/turma/{id}":{"get":{"tags":["Turma"],"summary":"Find by ID","description":"Returns details of an item by ID","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Turma"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"delete":{"tags":["Turma"],"summary":"Delete","description":"Deletes an item","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso"},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/turma/":{"get":{"tags":["Turma"],"summary":"List all","description":"Returns a list of all items","operationId":"findAll_2","responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Turma"}}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/professor/{id}":{"get":{"tags":["Professor"],"summary":"Find by ID","description":"Returns details of an item by ID","operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Professor"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"delete":{"tags":["Professor"],"summary":"Delete","description":"Deletes an item","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso"},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/professor/":{"get":{"tags":["Professor"],"summary":"List all","description":"Returns a list of all items","operationId":"findAll_3","responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Professor"}}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/pagamento/{id}":{"get":{"tags":["Pagamento"],"summary":"Find by ID","description":"Returns details of an item by ID","operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pagamento"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/aluno/{id}":{"get":{"tags":["Aluno"],"summary":"Find by ID","description":"Returns details of an item by ID","operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AlunoDto"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}},"delete":{"tags":["Aluno"],"summary":"Delete","description":"Deletes an item","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}},"/api/aluno/find/{value}":{"get":{"tags":["Aluno"],"summary":"Find by name","description":"Returns details of an item by name","operationId":"findByNomeOrCPF","parameters":[{"name":"value","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Operação realizada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlunoDto"}}}}},"400":{"$ref":"#/components/responses/badRequest"},"401":{"$ref":"#/components/responses/unauthorized"},"404":{"$ref":"#/components/responses/notFound"},"500":{"$ref":"#/components/responses/internalError"}}}}},"components":{"schemas":{"Professor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"Turma":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"periodo":{"type":"string"},"diaDeAula":{"type":"array","items":{"type":"string"}},"horarioInicioDeAula":{"type":"string"},"horarioFimDeAula":{"type":"string"},"professores":{"type":"array","items":{"$ref":"#/components/schemas/Professor"}}}},"Pagamento":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tipoDeAluno":{"type":"string"},"tipoDeAlunoEnum":{"type":"string","enum":["INTEGRAL","BOLSISTA"]},"valorMensalidade":{"type":"number"}}},"Aluno":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"cpf":{"type":"string"},"sexo":{"type":"string"},"dataNascimento":{"type":"string"},"telefone":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"responsavel":{"type":"array","items":{"$ref":"#/components/schemas/Responsavel"}},"turmas":{"type":"array","items":{"$ref":"#/components/schemas/Turma"}}}},"Endereco":{"type":"object","properties":{"cep":{"type":"string"},"rua":{"type":"string"},"bairro":{"type":"string"},"cidade":{"type":"string"},"numeroCasa":{"type":"string"},"complemento":{"type":"string"}}},"Responsavel":{"type":"object","properties":{"nome":{"type":"string"},"telefone":{"type":"string"},"email":{"type":"string"}}},"User":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"}}},"UserLogin":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}},"AlunoDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"cpf":{"type":"string"},"sexo":{"type":"string"},"dataNascimento":{"type":"string"},"telefone":{"type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"responsavel":{"type":"array","items":{"$ref":"#/components/schemas/Responsavel"}},"turmas":{"type":"array","items":{"$ref":"#/components/schemas/Turma"}},"pagamento":{"$ref":"#/components/schemas/Pagamento"}}}},"responses":{"badRequest":{"description":"BAD REQUEST"},"unauthorized":{"description":"UNAUTHORIZED"},"notFound":{"description":"NOT FOUND"},"internalError":{"description":"INTERNAL ERROR"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}